backend/client.py
"""
Gemini API クライアント
"""
from google import genai
from backend.config import GEMINI_API_KEY

# グローバルクライアントインスタンス
client = genai.Client(api_key=GEMINI_API_KEY)

