pyproject.toml
[project]
name = "gemini-api-backend"
version = "0.1.0"
description = "Gemini API機能を網羅したFastAPIバックエンド"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=0.2.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

